:lips: UI improvements
