+++ assets/style/scss/app.scss
@@ -21,3 +21,4 @@
 @import 'shared/header';
 @import 'shared/page';
 @import 'shared/persons';
+@import 'shared/card';
