+++ assets/style/scss/shared/_card.scss
@@ -1,0 +1,5 @@
+.app-card {
+  background-color: var(--color-ui-02);
+  border-radius: 0;
+  box-shadow: var(--shadow-01);
+}
